Refactored MOB cform conversion logic to util
