+++ app/pages/Settings/ChangePin.view/ChangePin.view.tsx
@@ -22,6 +22,10 @@ import { MOBIcon } from '../../../components/icons';
 import { PIN_MIN_SIZE } from '../../../constants/codes';
 import useIsMountedRef from '../../../hooks/useIsMountedRef';
 import type { Theme } from '../../../theme';
+import {
+  convertMobStringToPicoMobString,
+  convertPicoMobStringToMob,
+} from '../../../utils/convertMob';
 import isValidPin from '../../../utils/isValidPin';
 import { ChangePinViewProps } from './ChangePin';
 
@@ -91,39 +95,6 @@ const ChangePinView: FC<ChangePinViewProps> = ({
 
   const handleSelect = (event: ChangeEvent<HTMLInputElement>) => {
     event.target.select();
-  };
-
-  // TODO - ya, this definitely shouldn't live here
-  const PICO_MOB_PRECISION = 12;
-
-  const ensureMobStringPrecision = (mobString: string): string => {
-    const num = Number(mobString);
-    if (Number.isNaN(num)) {
-      throw new Error('mobString is NaN');
-    }
-
-    return num.toFixed(PICO_MOB_PRECISION);
-  };
-
-  // FIX-ME: This logic should not live here.
-  // Right now, we are aggressively assuming the number format is US local.
-  // We should have a universal solution to this problem
-  // Likely a component similiar to MOBNumberFormat with the ability to get the
-  // picoMob string (StringUInt64) value as a ref
-  const convertMobStringToPicoMobString = (mobString: string): string =>
-    ensureMobStringPrecision(mobString).replace('.', '');
-
-  // FIX-ME - seriously should not live here!
-  const convertPicoMobStringToMob = (picoMobString: string): string => {
-    if (picoMobString.length <= 12) {
-      return `0.${'0'.repeat(12 - picoMobString.length)}${picoMobString}`;
-    }
-
-    return [
-      picoMobString.slice(0, picoMobString.length - 12),
-      '.',
-      picoMobString.slice(picoMobString.length - 12),
-    ].join('');
   };
 
   return (
